+++ examples/cdftransport.py
@@ -5,35 +5,36 @@ if 'ccdfmean' not in cscripts :
     print("CDFtools not available")
     exit(0)
 
-# Declare "NEMO" project, both for Nemo raw outputs and for Monitoring outputs
+# Declare "data_CNRM" project, both for Nemo raw outputs and for Monitoring outputs
 # (with VT-files : mean values of vt, vs, ut, us for heat and salt transport)
 #
-cproject('NEMO')
+cproject('data_CNRM')
 
 # For 'standard' Nemo output files (actually, they are easier accessible using project "EM")
-root1="/cnrm/aster/data3/aster/senesi/NO_SAVE/expes/PRE6/${simulation}/O/"
+#root1="/cnrm/aster/data3/aster/senesi/NO_SAVE/expes/PRE6/${simulation}/O/"
+root1="/cnrm/aster/data1/UTILS/climaf/test_data/${simulation}/O/"
 suffix="${simulation}_1m_YYYYMMDD_YYYYMMDD_${variable}.nc"
 url_nemo_standard=root1+suffix  
 
 # For VT files from Monitoring
 #root2="/cnrm/aster/data3/aster/chevalli/Monitoring/Results/NO_SAVE/PRE6/SORTIE/PRE6/${simulation}/MONITOR/VT/"
-root2="/cnrm/aster/data1/UTILS/climaf/test_data/Monitoring/Results/${simulation}/MONITOR/VT/"
+root2="/cnrm/aster/data1/UTILS/climaf/test_data/${simulation}/VT/"
 url_nemo_monitoring=root2+suffix
 #
-dataloc(project='NEMO', organization='generic', url=[url_nemo_standard,url_nemo_monitoring])
+dataloc(project='data_CNRM', organization='generic', url=[url_nemo_standard,url_nemo_monitoring])
 # 
 # Declare how variables are scattered/grouped among files
 # (and with mixed variable names conventions - CNRM and  MONITORING)
-calias("NEMO","uo",filenameVar="grid_U_table2.3")
-calias("NEMO","vo",filenameVar="grid_V_table2.3")
-calias("NEMO","so,thetao",filenameVar="grid_T_table2.2")
+calias("data_CNRM","uo",filenameVar="grid_U_table2.3")
+calias("data_CNRM","vo",filenameVar="grid_V_table2.3")
+calias("data_CNRM","so,thetao",filenameVar="grid_T_table2.2")
 
 # Declare a special variable, composed of variables grouped in a file
 products="vomevt,vomevs,vozout,vozous"
-calias("NEMO",products,filenameVar="VT")
+calias("data_CNRM",products,filenameVar="VT")
 
 # Define defaults facets for datasets 
-cdef("project","NEMO")
+cdef("project","data_CNRM")
 cdef("frequency","monthly")
 cdef("simulation","PRE6CPLCr2alb")
 cdef("period","199808-199809")
@@ -44,9 +45,9 @@ dvo=ds(variable="vo")
 dx=ds(variable=products)
 
 # Tell how to bring required fixed files to cdftransport
-# (this can use wildcards ${model}, ${project}, ${simulation})
+# (this can use wildcards ${model}, ${project}, ${simulation}, ${realm})
 #tpath='/cnrm/aster/data3/aster/chevalli/Monitoring/MONITORING_v3.1/config/'
-tpath='/cnrm/aster/data1/UTILS/climaf/test_data/Monitoring/MONITORING_v3.1/config/'
+tpath='/cnrm/aster/data1/UTILS/climaf/test_data/fixed/'
 fixed_fields('ccdftransport',
              ('mesh_hgr.nc',tpath+'ORCA1_mesh_hgr.nc'),
              ('mesh_zgr.nc',tpath+'ORCA1_mesh_zgr.nc'))